@NonNull and @Nullable added, for Kotlin interoperability.

+++ mobile/src/main/java/io/syslogic/github/model/RateLimit.java
@@ -2,8 +2,6 @@ package io.syslogic.github.model;
 
 import com.google.gson.annotations.SerializedName;
 
-import java.util.ArrayList;
-
 public class RateLimit {
 
     @SerializedName("limit")
@@ -15,7 +13,7 @@ public class RateLimit {
     @SerializedName("reset")
     private long reset;
 
-    public long getLimit() {
+    public int getLimit() {
         return limit;
     }
